Moves java test classes from 'common' -> 'se'
